+++ stories/components/notice-bubble/docs/examples/dynamic_notice.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { NoticeBubbleContainer, NoticeBubbleManager } from 'intergalactic/notice-bubble';
-import Button from 'intergalactic/button';
-import WarningM from 'intergalactic/icon/Warning/m';
-import ReloadM from 'intergalactic/icon/Reload/m';
-import Spin from 'intergalactic/spin';
-import { Flex } from 'intergalactic/flex-box';
+import { NoticeBubbleContainer, NoticeBubbleManager } from '@semcore/notice-bubble';
+import Button from '@semcore/button';
+import WarningM from '@semcore/icon/Warning/m';
+import ReloadM from '@semcore/icon/Reload/m';
+import Spin from '@semcore/spin';
+import { Flex } from '@semcore/flex-box';
 
 const manager = new NoticeBubbleManager();