+++ semcore/notice-bubble/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.48.0] - 2024-11-21
+
+### Added
+
+- `replaceLast` method to the Manager.
+
+### Fixed
+
+- Warning about `Hint` title.
+
 ## [5.47.1] - 2024-11-08
 
 ### Changed